LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-showModal-expected.txt


PASS dialog element: showModal()
PASS showModal() on a <dialog> that already has an open attribute throws an InvalidStateError exception
PASS showModal() on a <dialog> after initial showModal() and removing the open attribute
PASS showModal() on a <dialog> not in a Document throws an InvalidStateError exception
PASS when opening multiple dialogs, only the newest one is non-inert
PASS opening dialog without focusable children
PASS opening dialog with multiple focusable children
FAIL opening dialog with multiple focusable children, one having the autofocus attribute assert_equals: expected Element node <input id="i82" value="foobar" autofocus=""></input> but got Element node <body><div id="log"></div>
<button id="b0">OK</button>
<d...
PASS when opening multiple dialogs, the most recently opened is rendered on top
OK
